refactor(auth): rename databaseClock to sessionExpiration
