+++ web/src/features/auth-credentials/lib/credentialsServerUtils.ts
@@ -1,6 +1,6 @@
 import { createHash } from "crypto";
 import { createProjectMembershipsOnSignup } from "@/src/features/auth/lib/createProjectMembershipsOnSignup";
-import { advanceSessionsExpiredAtForEmail } from "@/src/features/auth/lib/databaseClock";
+import { advanceSessionsExpiredAtForEmail } from "@/src/features/auth/lib/sessionExpiration";
 import type { AdClickIds } from "@/src/features/auth";
 import { env } from "@/src/env.mjs";
 import { prisma } from "@langfuse/shared/src/db";
